src/common/common_requirements.txt
grpcio==1.56.2
protobuf==4.23.4
pyyaml==6.0.1
python-jose==3.3.0
pycryptodome==3.20.0